requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements.in/dev.txt
#
appdirs==1.4.3            # via black
attrs==19.3.0             # via black
black==19.10b0
blinker==1.4              # via nplusone
click==7.1.1              # via black, pip-tools
django-admin-honeypot==1.1.0
django-debug-toolbar==2.2
django-js-asset==1.2.2    # via django-mptt
django-mptt==0.10.0
django-widget-tweaks==1.4.8
django==2.2.5
html5lib==1.0.1           # via xhtml2pdf
isort==4.3.21
nplusone==1.0.0
pathspec==0.7.0           # via black
pdf2image==1.10.0
pillow==7.0.0             # via pdf2image, reportlab, xhtml2pdf
pip-tools==4.1.0
psycopg2-binary==2.8.3
pypdf2==1.26.0            # via xhtml2pdf
pytz==2019.3              # via django
regex==2020.2.20          # via black
reportlab==3.5.42         # via xhtml2pdf
six==1.14.0               # via html5lib, nplusone, pip-tools, xhtml2pdf
sqlparse==0.3.1           # via django, django-debug-toolbar
toml==0.10.0              # via black
typed-ast==1.4.1          # via black
webencodings==0.5.1       # via html5lib
xhtml2pdf==0.2.3
